New Updates:  - Updated source to be more consistent with Crystal standards.  - Minor fixes to some logic to prevent some unindended issues.  - Added a couple new hashes to the massive list of hash -> plaintext.  - Improved the regex scan used to find plaintext file names to get a few more of those stragglers.  - Added some data output at the end to show how many file names were dehashed and how many weren't.

+++ src/util/c-src/fnv64.c
@@ -1,6 +1,6 @@
 #define ull unsigned long long
 
-ull fnv64(const char* string) {
+unsigned long long fnv64(const char* string) {
     ull hash = 0x47F5817A5EF961BA;
     ull prime = 0x100000001B3;
